refactor(onboarding): default Content() for onboarding blocs
Move all seven onboarding step screens (Welcome, StartCooking,
SaveRecipes, MealPlanning, CookMode, GroceryList, OnboardingRoot) plus
the shared OnboardingInfoLayout helper into the public module. Each
onboarding Bloc now provides a default Content() that delegates to its
screen; impl Blocs and OnboardingPreviews drop their Content overrides.
