+++ client/onboarding/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/onboarding/impl/SaveRecipesBlocImpl.kt
@@ -1,12 +1,9 @@
 package com.plusmobileapps.chefmate.onboarding.impl
 
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.di.AppScope
 import com.plusmobileapps.chefmate.onboarding.SaveRecipesBloc
-import com.plusmobileapps.chefmate.onboarding.impl.ui.SaveRecipesScreen
 import com.plusmobileapps.metro.extensions.assistedfactory.ContributesAssistedFactory
 import dev.zacsweers.metro.Assisted
 import dev.zacsweers.metro.AssistedInject
@@ -23,10 +20,5 @@ class SaveRecipesBlocImpl(
 
     override fun onNextClicked() {
         output.onNext(SaveRecipesBloc.Output.Next)
-    }
-
-    @Composable
-    override fun Content(modifier: Modifier) {
-        SaveRecipesScreen(bloc = this, modifier = modifier)
     }
 }
