+++ client/onboarding/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/onboarding/impl/WelcomeBlocImpl.kt
@@ -1,12 +1,9 @@
 package com.plusmobileapps.chefmate.onboarding.impl
 
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.di.AppScope
 import com.plusmobileapps.chefmate.onboarding.WelcomeBloc
-import com.plusmobileapps.chefmate.onboarding.impl.ui.WelcomeScreen
 import com.plusmobileapps.metro.extensions.assistedfactory.ContributesAssistedFactory
 import dev.zacsweers.metro.Assisted
 import dev.zacsweers.metro.AssistedInject
@@ -28,10 +25,5 @@ class WelcomeBlocImpl(
 
     override fun onSkipClicked() {
         output.onNext(WelcomeBloc.Output.Skip)
-    }
-
-    @Composable
-    override fun Content(modifier: Modifier) {
-        WelcomeScreen(bloc = this, modifier = modifier)
     }
 }
