+++ client/onboarding/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/onboarding/CookModeBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.onboarding
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.ui.ComposeScreen
@@ -7,6 +9,11 @@ import com.plusmobileapps.chefmate.ui.ComposeScreen
 /** Onboarding step explaining how to start Cook Mode from the recipe detail screen. */
 interface CookModeBloc : ComposeScreen {
     fun onNextClicked()
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        CookModeScreen(bloc = this, modifier = modifier)
+    }
 
     sealed class Output {
         /** Advance to the next onboarding step. */
