+++ client/onboarding/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/onboarding/MealPlanningBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.onboarding
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.ui.ComposeScreen
@@ -7,6 +9,11 @@ import com.plusmobileapps.chefmate.ui.ComposeScreen
 /** Onboarding step explaining how to add recipes to meals across the days of the week. */
 interface MealPlanningBloc : ComposeScreen {
     fun onNextClicked()
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        MealPlanningScreen(bloc = this, modifier = modifier)
+    }
 
     sealed class Output {
         /** Advance to the next onboarding step. */
