+++ client/onboarding/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/onboarding/OnboardingRootBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.onboarding
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.arkivanov.decompose.router.stack.ChildStack
 import com.arkivanov.decompose.value.Value
 import com.arkivanov.essenty.backhandler.BackHandlerOwner
@@ -20,6 +22,11 @@ import com.plusmobileapps.chefmate.ui.ComposeScreen
  */
 interface OnboardingRootBloc : BackHandlerOwner, BackClickBloc, ComposeScreen {
     val routerState: Value<ChildStack<*, Child>>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        OnboardingRootScreen(bloc = this, modifier = modifier)
+    }
 
     sealed class Child {
