+++ client/onboarding/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/onboarding/StartCookingBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.onboarding
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.ui.ComposeScreen
@@ -7,6 +9,11 @@ import com.plusmobileapps.chefmate.ui.ComposeScreen
 /** Final screen of the onboarding flow — confirms setup and launches the rest of the app. */
 interface StartCookingBloc : ComposeScreen {
     fun onStartCookingClicked()
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        StartCookingScreen(bloc = this, modifier = modifier)
+    }
 
     sealed class Output {
         /** The user is done with onboarding and wants to start using the app. */
