+++ client/onboarding/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/onboarding/WelcomeBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.onboarding
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.ui.ComposeScreen
@@ -7,6 +9,11 @@ import com.plusmobileapps.chefmate.ui.ComposeScreen
 /** First screen of the onboarding flow — greets the user and kicks off the flow. */
 interface WelcomeBloc : ComposeScreen {
     fun onGetStartedClicked()
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        WelcomeScreen(bloc = this, modifier = modifier)
+    }
 
     fun onSignInClicked()
